Time-Time and Objective-Objective Relationships Omitted in Fingerprints
In the performance fingerprints used for clustering, we try to fit a
model to the algorithm behavior. The parameters of the model form the
fingerprint. Now we just model time-objective relationships and omit
time-time and objective-objective relationships, which may occur if we
have multiple time or objective dimensions. These relationships may be
reconstructed from the time-objective ones if necessary anyway, so we
can safe this computational effort.