First Steps towards Transparent Parallelism
The JobExecutor class is implemented in different flavors, either for
sequential or parallel job execution with and without waiting for the
jobs to finish. The implementations may detect ForkJoinPools and try
to submit the jobs to them if the selected form of parallelism allows
it. This will allow us to parallelize tasks in our computations in a
transparent way, without need to deal with ForkJoinPools or -Tasks at
all, while still reaping their benefits.