Improved Execute and Fitting Examples
The class Execute has been improved (or bug-fixed, if you want). It will
now only use a ForkJoinPool to execute jobs if the calling task already
runs in such a pool. Otherwise it will execute jobs directly. This has
several reasons, such as: It is sort of awkward to leak tasks into the
common ForkJoinPool in Java 8, as some of our tasks may take very long
and thus may stall other things going on in said pool. Moreover, if we
want to let our jobs creep into a ForkJoinPool while not being already
in one, we may fail if the pool is currently shutting down. Finally,
if we always use the pool's submit function to mark jobs for execution,
we may again fail if the pool is shutting down, even if we legally try
to run the tasks if we are already in the pool. Thus, I don't do this
anymore.

The function fitting example has been improved to provide better
statistics and to run its jobs in parallel.