Code Generators now in `test` Folder; Better Add3 and Add4
I moved the code generators from the `src` to the `test` folder, as
they should not be included in the produced jar archives. This makes
the jars smaller.

I developed a code generator for high-precision adding. This code is
based on Kahan sums, which are more numerically stable than just adding
up three numbers. In order to achieve maximum precision, we not just
sum up the n numbers, we instead compute the Kahan sum of each of the n!
permutations of the numbers and return the one sum with smallest
absolute compensation.

See https://en.wikipedia.org/wiki/Kahan_summation_algorithm