Much Better Way to Add N Numbers + BugFix for Tests
Based on the Python sources, I was able to implement a much more stable
and accurate way to sum up 3, 4, or n numbers. The n numbers are stored
in an array of length n which, during the process of computing the sum,
is overwritten with compensation registers. The use of at most n of
these allows the sum to become accurate.

http://stackoverflow.com/users/2955095/
http://code.activestate.com/recipes/393090-binary-floating-point-summation-accurate-to-full-p/
https://www.python.org/getit/source/ (Python 3.5.1rc1 - 2015-11-23)

The MathematicalFunctionTest now actually tests all permutations of all
input parameters whenever appropriate. As it should.