Improved Execute, ArrayIterator, and FittingExamples
Execute now uses the same code to join jobs regardless of whether the
jobs are stored in an (Iterable) list or array and regardless of
whether we want to obtain the results or discard them directly.

ArrayIterator now also implements Iterable. Its iterate() method always
returns 'this'. The idea is that we can now also use this iterator
directly in the new 'for(T x : Iterable<T>)'-style loops directly. This
is done, for example, in the above Execute class.

The FittingExamples now provide some logging information, so that we
can see what the progress is.