Improved Function Fitter and Execute
The function fitting job based on least squares now produces relatively
high-quality and stable results. It will try to repeat its execution
several times if the results seem to be unstable in order to become
more robust. Of course, this cannot guide against getting locally-
optimal results, but this is as good as I can do.

Execute now can send jobs to the common Fork-Join Pool. Such a pool
exists under Java 8. If it cannot be found (maybe because we are under
Java 7), such a pool will be created internally.