Improved Function Fitting & Refactoring of Package Structure
Function fitting is now much more reliable (and a bit slower). Its
results are now much more reproducible (and better).

The package structure for function fitting has been redesigned: It is
now part of a "ml" package instead of "math". The reason is that I will
add more Machine Learning functionality, such as clustering and
classification.