Machine Learning Utilities now support Clustering
The machine learning utilities package has been extended with clustering
functionality. This functionality is specified as a set of interfaces.
A set of basic classes implements these interfaces without actual
clustering functionality but with some data and cluster normalization
ability. The R script used for the instance finger print clustering in
the experiment data evaluation package has been adapted and moved to as
first full clusterer implementation into the Machine Learning package.
It thus can now be used in a variety of different scenarios without
requiring any foot work or understanding of the interaction with R.