New Attribute: Modeling of Dimension Relationships on Instance Run Sets
A new attribute has been introduced which models the relationships of
different measure dimensions over all runs of a given instance run set.
It takes two dimensions, an input and an output dimension, and computes
a model which describes their relationship. For this purpose, multiple
models are fitted and tested subsequently and the best one is picked.

The new attribute comes in two flavors: In its plain form, called
DimensionRelationship, it just returns the model, its parameters, and
the model quality as IFittingResult record. In its extended form, it
returns the IFittingQualityMeasure used for the fitting as well. The
quality measure takes quite some space in memory (it holds a big matrix)
and building it is a bit expensive (but not much). So, in computations
where we may need both the fitting and the quality fitting (such as
when we want to compare how other models fit to the same data), we may
use this attribute.

Both attributes only (and permanently) preserve the fitting, the quality
measure is never stored.