Clustering now also supports Distance Matrices
The clustering API in the Machine Learning utilities package was
designed to deal only with m*n data matrices, where each of the m rows
represents one data sample characterized by n feature values.

The API now also supports clustering by distance matrices: Here, the
data itself is not used, but a m*m matrix specifying the distances
between the m data samples. This matrix will usually have all 0s at the
diagonal and the values in the triangle above the diagonal is mirror
those in below the diagonal, i.e., the distance of one sample to itself
is 0 and the distance from sample A to sample B should be the same as
the distance from sample B to sample A.

Classes to support this behavior have been introduced. This led to an
improvement of the documentation of the current matrix classes.

Furthermore (and unrelated to that), the examples for the function
fitting now allow for multiple models being fitted and the best one
chosen. This is to facilitate using different models and trying to find
the best one for a given data set. This makes sense, since we cannot
predict the one, single best model for algorithm performance (in our
experimentation API). Instead, we may have a few different models which
come close to what we observed in practice and let the fitter decide
which one fits best.